Add section postfix to other tabs
